refactor: Merge Linux and Mac script
