--- src/main/groovy/ru/endlesscode/bukkitgradle/server/LinuxScript.groovy
@@ -1,27 +1,0 @@
-package ru.endlesscode.bukkitgradle.server
-
-import ru.endlesscode.bukkitgradle.extension.RunConfiguration
-
-class LinuxScript extends SystemScript {
-    LinuxScript(RunConfiguration configuration, String version) {
-        super(configuration, version)
-    }
-
-    @Override
-    protected String getExt() {
-        return "sh"
-    }
-
-    @Override
-    protected String getScriptText() {
-        """#!/bin/sh
-
-${buildRunCommand()}"""
-    }
-
-    @Override
-    List<String> getCommand() {
-        String scriptName = getFileName()
-        return ["sh", scriptName]
-    }
-}
